refactor: Mock Account Repository to be used for other tests
