+++ src/usecase/account.rs
@@ -29,42 +29,10 @@ pub fn delete_account(
 mod tests {
     use super::*;
     use crate::domain::object::account::AccountId;
+    use crate::tests::mock_account_repository::MockAccountRepository;
     use chrono::{Duration, Local};
     use std::cell::RefCell;
     use std::collections::HashMap;
-
-    use anyhow::Result;
-
-    struct MockAccountRepository {
-        pool: RefCell<HashMap<i64, Account>>,
-    }
-
-    impl AccountRepository for MockAccountRepository {
-        fn insert(&self, account: &Account) -> Result<()> {
-            let _ = &self
-                .pool
-                .borrow_mut()
-                .entry(account.id.get())
-                .or_insert_with(|| account.clone());
-
-            Ok(())
-        }
-        fn list(&self) -> Result<Vec<Account>> {
-            let accounts: Vec<Account> = self.pool.borrow().values().cloned().collect();
-            Ok(accounts)
-        }
-
-        fn find_by_id(&self, account_id: &AccountId) -> Result<Account> {
-            match self.pool.borrow().get(&account_id.get()) {
-                Some(account) => Ok(account.clone()),
-                None => Err(anyhow::anyhow!("Account not found")),
-            }
-        }
-        fn delete(&self, account: &Account) -> Result<()> {
-            let _ = &self.pool.borrow_mut().remove(&account.id.get());
-            Ok(())
-        }
-    }
 
     #[test]
     fn success_post_account() {
